+++ needle/core/framework/framework.py
@@ -3,6 +3,7 @@ import os
 import sys
 import cmd
 import codecs
+import readline
 import traceback
 
 from options import Options
@@ -139,6 +140,24 @@ class Framework(cmd.Cmd):
         options = sorted(self._get_show_names())
         return [x for x in options if x.startswith(text)]
 
+    def _history_save(self):
+        history_path = Constants.FILE_HISTORY
+        try:
+            if self._global_options['save_history']:
+                self.printer.debug("Saving command history to: {}".format(history_path))
+                readline.write_history_file(history_path)
+        except Exception as e:
+            self.printer.warning("Error while saving command history: {}".format(e))
+            self.printer.warning("Continuing anyway...")
+
+    def _history_load(self):
+        history_path = Constants.FILE_HISTORY
+        if os.path.exists(history_path):
+            self.printer.debug("Trying to load command history from: {}".format(history_path))
+            readline.read_history_file(history_path)
+        else:
+            self.printer.debug("Command history not found in: {}".format(history_path))
+
     # ==================================================================================================================
     # OUTPUT METHODS
     # ==================================================================================================================
@@ -253,11 +272,12 @@ class Framework(cmd.Cmd):
             print(pattern % ('Name'.ljust(key_len), 'Current Value'.ljust(val_len), 'Required', 'Description'))
             print(pattern % (self.ruler*key_len, (self.ruler*13).ljust(val_len), self.ruler*8, self.ruler*11))
             for key in sorted(options):
-                value = options[key] if options[key] != None else ''
-                reqd = 'no' if options.required[key] is False else 'yes'
-                desc = options.description[key]
-                print(pattern % (key.upper().ljust(key_len), Utils.to_unicode_str(value).ljust(val_len),
-                                 Utils.to_unicode_str(reqd).ljust(8), desc))
+                if not key == Constants.PASSWORD_CLEAR:
+                    value = options[key] if options[key] != None else ''
+                    reqd = 'no' if options.required[key] is False else 'yes'
+                    desc = options.description[key]
+                    print(pattern % (key.upper().ljust(key_len), Utils.to_unicode_str(value).ljust(val_len),
+                                     Utils.to_unicode_str(reqd).ljust(8), desc))
             print('')
         else:
             print('')
@@ -344,68 +364,21 @@ class Framework(cmd.Cmd):
             # if value type is bool or int, then we know the options is set
             if not type(self.options[option]) in [bool, int]:
                 if self.options.required[option] is True and not self.options[option]:
+                    if option == Constants.PASSWORD_CLEAR:
+                        option = 'password'.upper()
                     raise FrameworkException('Value required for the \'%s\' option.' % (option.upper()))
         return
 
     def register_option(self, name, value, required, description):
         self.options.init_option(name=name.lower(), value=value, required=required, description=description)
-        # TODO: support for config file
-        #self._load_config()
-
-    # TODO: support for config file
-    '''
-    def _load_config(self):
-        config_path = os.path.join(self._home, 'config.dat')
-        # don't bother loading if a config file doesn't exist
-        if os.path.exists(config_path):
-            # retrieve saved config data
-            with open(config_path) as config_file:
-                try:
-                    config_data = json.loads(config_file.read())
-                except ValueError:
-                    # file is corrupt, nothing to load, exit gracefully
-                    pass
-                else:
-                    # set option values
-                    for key in self.options:
-                        try:
-                            self.options[key] = config_data[self._modulename][key]
-                        except KeyError:
-                            # invalid key, contnue to load valid keys
-                            continue
-
-    def _save_config(self, name):
-        config_path = os.path.join(self._home, 'config.dat')
-        # create a config file if one doesn't exist
-        open(config_path, 'a').close()
-        # retrieve saved config data
-        with open(config_path) as config_file:
-            try:
-                config_data = json.loads(config_file.read())
-            except ValueError:
-                # file is empty or corrupt, nothing to load
-                config_data = {}
-        # create a container for the current module
-        if self._modulename not in config_data:
-            config_data[self._modulename] = {}
-        # set the new option value in the config
-        config_data[self._modulename][name] = self.options[name]
-        # remove the option if it has been unset
-        if config_data[self._modulename][name] is None:
-            del config_data[self._modulename][name]
-        # remove the module container if it is empty
-        if not config_data[self._modulename]:
-            del config_data[self._modulename]
-        # write the new config data to the config file
-        with open(config_path, 'w') as config_file:
-            json.dump(config_data, config_file, indent=4)
-    '''
 
     # ==================================================================================================================
     # COMMAND METHODS
     # ==================================================================================================================
     def do_exit(self, params):
         """Stop background jobs, cleanup temp folders (local&remote), close connection, then exits the Framework."""
+        # Save history
+        self._history_save()
         # Stop background jobs
         for i in xrange(len(self._jobs)):
             self.do_kill(i)
@@ -453,8 +426,15 @@ class Framework(cmd.Cmd):
         name = options[0].lower()
         if name in self.options:
             value = ' '.join(options[1:])
+
+            if name == 'password':
+                self.options[Constants.PASSWORD_CLEAR] = value
+                value = Constants.PASSWORD_MASK
+
+            # Actual set
             self.options[name] = value
             print('%s => %s' % (name.upper(), value))
+
             # Check verbosity level
             if name == 'debug':
                 self.printer.set_debug(self.options['debug'])
@@ -468,8 +448,6 @@ class Framework(cmd.Cmd):
                 self.printer.debug("Output folder changed, reloading modules")
                 self._local_ready = Framework._local_ready = False
                 self.do_reload(None)
-            # TODO: support for config file
-            #self._save_config(name)
         else:
             self.printer.error('Invalid option.')
 
@@ -626,7 +604,7 @@ class Framework(cmd.Cmd):
         IP = self._global_options['ip']
         PORT = self._global_options['port']
         USERNAME = self._global_options['username']
-        PASSWORD = self._global_options['password']
+        PASSWORD = self._global_options[Constants.PASSWORD_CLEAR]
         PUB_KEY_AUTH = self._global_options['pub_key_auth']
         return IP, PORT, USERNAME, PASSWORD, PUB_KEY_AUTH
 
@@ -660,7 +638,7 @@ class Framework(cmd.Cmd):
             if self._global_options['ip'] != self.device._ip or \
                self._global_options['port'] != self.device._port or \
                self._global_options['username'] != self.device._username or \
-               self._global_options['password'] != self.device._password or \
+               self._global_options[Constants.PASSWORD_CLEAR] != self.device._password or \
                self._global_options['pub_key_auth'] != self.device._pub_key_auth:
 
                 self.printer.verbose('Settings changed in global options. Establishing a new connection')